requirements.txt
dipy>=1.6.0
matplotlib>=3.7.0
nibabel>=5.0.1
notebook>=6.5.1
numba>=0.55.1
numpy>=1.24.1
pandas>=2.0.0
pathos>=0.3.0
renalsegmentor>=1.3.5
scikit-image>=0.20.0, <0.22
scipy>=1.9.1, <1.12
scikit-learn~=1.2.1  # Changing versions is known to cause changes in iSNR results
tabulate>=0.9.0
tqdm>=4.64.1
SimpleITK-SimpleElastix>=2.0.0rc2.dev910
